More refactoring of AWSSQLLiteNormalizedCache.swift; More Promise wrapping around AWSSQLLiteNormalizedCache
