+++ AWSAppSyncClient/AWSAppSyncCache.swift
@@ -1,0 +1,28 @@
+//
+// Copyright 2010-2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License").
+// You may not use this file except in compliance with the License.
+// A copy of the License is located at
+//
+// http://aws.amazon.com/apache2.0
+//
+// or in the "license" file accompanying this file. This file is distributed
+// on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+// express or implied. See the License for the specific language governing
+// permissions and limitations under the License.
+//
+
+import Foundation
+
+public protocol MutationCache {
+    func saveMutation(body: Data) -> Int64
+    func getMutation(id: Int64) -> Data
+    func loadAllMutation() -> [Int64: Data]
+}
+
+public enum AWSSQLLiteNormalizedCacheError: Error {
+    case invalidRecordEncoding(record: String)
+    case invalidRecordShape(object: Any)
+    case invalidRecordValue(value: Any)
+}
